+++ hvega/CHANGELOG.md
@@ -13,6 +13,9 @@ The `Operation` type has gained the `Product` specifier from Vega-Lite
 
 The `TextChannel` has gained `TStrings` to support multi-line labels.
 
+The `VAlign` type has gained `AlignLineTop` and `AlignLineBottom`
+(Vega-Lite 4.6.0).
+
 ## 0.6.0.0
 
 The Vega-Lite tests are now validated against version 4.5 of the